+++ R/plot-progression-table.R
@@ -7,7 +7,8 @@
 #' @param adjustment_multiplier Factor to multiply the adjustment. Useful when converting to percentage.
 #'     Default is 1
 #' @param plot Character string. Options include "%1RM" (default) and "adjustment"
-#' @param label_size Numeric. Default is 3
+#' @param font_size Numeric. Default is 8
+#' @param label_size Numeric. Default is 2.5
 #' @param ... Forwarder to the \code{\link{generate_progression_table}} function
 #' @return \code{ggplot2} object
 #' @export
@@ -25,7 +26,8 @@ plot_progression_table <- function(progression_table = progression_RIR_increment
                                    plot = "%1RM",
                                    signif_digits = 3,
                                    adjustment_multiplier = 1,
-                                   label_size = 3,
+                                   font_size = 8,
+                                   label_size = 2.5,
                                    ...) {
 
   # +++++++++++++++++++++++++++++++++++++++++++
@@ -61,7 +63,7 @@ plot_progression_table <- function(progression_table = progression_RIR_increment
 
   gg <- switch(plot,
     "%1RM" = ggplot2::ggplot(progression_tbl, ggplot2::aes(x = step, y = reps)) +
-      ggplot2::theme_linedraw() +
+      ggplot2::theme_linedraw(font_size) +
       # geom_tile(fill = "transparent", color = "black") +
       ggplot2::geom_text(ggplot2::aes(label = perc_1RM), size = label_size) +
       ggplot2::scale_y_discrete(limits = rev(levels(progression_tbl$reps))) +